+++ android/app/src/main/java/com/app/belcobtm/domain/wallet/item/CoinFeeDataItem.kt
@@ -2,7 +2,7 @@ package com.app.belcobtm.domain.wallet.item
 
 class CoinFeeDataItem(
     val txFee: Double,
-    val byteFee: Double,
+    val byteFee: Long,
     val recallFee: Double?,
     val gasPrice: Double,
     val gasLimit: Double,
